Support override RenderingScale calling NewFrame
Found from testing  JUCE's openGL rendering scale can be inaccurate on certain displays and platforms.
E.g. running windows via bootcamp the renderingScale is set to 1, but the display dpi is 144 and the display scale is 1, and so the rendering scale should be calculated as:

renderingScale = displayScale * (displayDPI / platformDPI)

So allowing users to override this if run into this issue
(PlatformDPI is e.g. 72 on Mac, 96 on Windows)